LayoutTests/media/audio-constructor-autobuffer.html

<html>
    <head>
        <script src=video-test.js></script>

        <script>

            function loaded()
            {
                consoleWrite("++ Test after 'load' event");
                testExpected("media.getAttribute('autobuffer')", null, "!=");
                testExpected("media.autobuffer", true);
                endTest();
            }

            function start()
            {
                run("media = new Audio()");
                consoleWrite("");

                consoleWrite("++ Test initial attribute value");
                testExpected("media.getAttribute('autobuffer')", null, "!=");
                testExpected("media.autobuffer", true);
                consoleWrite("");

                media.onload = loaded;
                run("media.src = 'content/test.wav'");
                run("media.load()");

                consoleWrite("");
            }

        </script>
    </head>

    <body onload="start()">
        <p>Test that Audio() sets 'autobuffer' attribute.</p>
    </body>
</html>
